+++ task/big_file_sort/main_sort_memlimit.cpp
@@ -212,7 +212,7 @@ int main(int argc, char* argv[])
     std::string::size_type sz;
     double limit_gb = std::stod(ssize, &sz);
     double k_safe = 0.975;
-    const size_t MEMORY_LIMIT = k_safe * limit_gb * GB_SIZE / hardware_concurrency;
+    const size_t MEMORY_LIMIT = static_cast<size_t>(k_safe * limit_gb * GB_SIZE / hardware_concurrency);
 
     size_t blob_counter = 0;
     cblobs blobs;
